fix incompatibilities with debian bullseye

+++ papers/pair-correlation/figs/plot-path-triplet-contact.py
@@ -140,7 +140,7 @@ g3dft = read_triplet(ff, 'this-work-short')
 zdft = loadtxt("figs/triplet-z.dat")
 xdft = loadtxt("figs/triplet-x.dat")
 
-levels = linspace(0, gmax, gmax*100)
+levels = linspace(0, gmax, int(gmax*100))
 gr = read_gr(ff)
 ginf = interp(rpath, gr[:, 0], gr[:, 1]/ff)
 xlo = 0.5*ginf/gmax
